Significant improvements to Signal.scan functionality
Iterative scanning of a signal items to update a state signal can now be
done over multiple item signals each with its own update function.
Scanning can also take one or more signal (or value) parameters which
get passed as the last args to the update functions. Added
sig.quiescenceWatch which is implemented with these features. This
update breaks existing scan functions as the item and state args are now
switched to be consistent with the order of Signal.scan